moved util function, ticket and download queue functionality to their own files the aim being to reduce global vars
