+++ source/utils2.cpp
@@ -1,0 +1,115 @@
+#include "utils2.h"
+
+u64 hex_to_u64( std::string value) {
+    u64 out;
+    std::istringstream(value) >> std::hex >> out;
+    return out;
+}
+
+std::string u32_to_hex_string(u32 i)
+{
+    std::stringstream stream;
+    stream << std::setfill ('0') << std::setw(sizeof(u32)*2) << std::hex << i;
+    return stream.str();
+}
+
+
+std::string ToHex(const std::string& s)
+{
+    std::ostringstream ret;
+    for (std::string::size_type i = 0; i < s.length(); ++i)
+    {
+        int z = s[i]&0xff;
+        ret << std::hex << std::setfill('0') << std::setw(2) << z;
+    }
+    return ret.str();
+}
+
+int mkpath(std::string s,mode_t mode)
+{
+    size_t pre=0,pos;
+    std::string dir;
+    int mdret = 0;
+
+    if(s[s.size()-1]!='/'){
+        // force trailing / so we can handle everything in loop
+        s+='/';
+    }
+
+    while((pos=s.find_first_of('/',pre))!=std::string::npos){
+        dir=s.substr(0,pos++);
+        pre=pos;
+        if(dir.size()==0) continue; // if leading / first time is 0 length
+        if((mdret=mkdir(dir.c_str(),mode)) && errno!=EEXIST){
+            return mdret;
+        }
+    }
+    return mdret;
+}
+
+char parse_hex(char c)
+{
+    if ('0' <= c && c <= '9') return c - '0';
+    if ('A' <= c && c <= 'F') return c - 'A' + 10;
+    if ('a' <= c && c <= 'f') return c - 'a' + 10;
+    std::abort();
+}
+
+char* parse_string(const std::string & s)
+{
+    char* buffer = new char[s.size() / 2];
+    for (std::size_t i = 0; i != s.size() / 2; ++i)
+        buffer[i] = 16 * parse_hex(s[2 * i]) + parse_hex(s[2 * i + 1]);
+    return buffer;
+}
+
+std::string get_file_contents(const char *filename)
+{
+  std::ifstream in(filename, std::ios::in | std::ios::binary);
+  if (in)
+  {
+    return(std::string((std::istreambuf_iterator<char>(in)), std::istreambuf_iterator<char>()));
+  }
+  throw(errno);
+}
+
+void removeForbiddenChar(std::string* s, bool onlyCR = false)
+{
+    std::string::iterator it;
+    std::string illegalChars = "\n";
+    if(!onlyCR) {
+        illegalChars = "\n\\/:?\"<>|";
+    }
+    for (it = s->begin() ; it < s->end() ; ++it){
+        bool found = illegalChars.find(*it) != std::string::npos;
+        if(found)
+        {
+            *it = ' ';
+        }
+    }
+}
+
+std::istream& GetLine(std::istream& is, std::string& t)
+{
+    t.clear();
+    std::istream::sentry se(is, true);
+    std::streambuf* sb = is.rdbuf();
+
+    for (;;) {
+        int c = sb->sbumpc();
+        switch (c) {
+            case '\n':
+              return is;
+            case '\r':
+              if (sb->sgetc() == '\n')
+                sb->sbumpc();
+              return is;
+            case  EOF:
+              if (t.empty())
+                is.setstate(std::ios::eofbit);
+              return is;
+            default:
+              t += (char)c;
+        }
+    }
+}
